[Loom] Isolate source lowering module state
Give pass-local module lowering state a documented component boundary
instead of keeping its representation and allocation logic in the
general lowering context.

Keep the representation private, retain the existing arena lifetime and
static-key interning contract, and add focused coverage for zero
initialization, identity, record growth, and raw storage allocation.
