+++ loom/src/loom/codegen/low/lower/BUILD.bazel
@@ -38,6 +38,7 @@ loom_cc_library(
         "lower_rule_source_memory_match.c",
         "lower_rule_source_memory_materialize.c",
         "lower_rules.c",
+        "module_state.c",
     ],
     hdrs = [
         "contract_query.h",
@@ -46,6 +47,7 @@ loom_cc_library(
         "lower_rule_descriptor.h",
         "lower_rule_source_memory.h",
         "lower_rules.h",
+        "module_state.h",
     ],
     visibility = _LOWER_VISIBILITY,
     deps = [
@@ -115,6 +117,17 @@ loom_cc_library(
         "//loom/src/loom/target:types",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/internal:arena",
+    ],
+)
+
+loom_cc_test(
+    name = "module_state_test",
+    srcs = ["module_state_test.cc"],
+    deps = [
+        ":lower",
+        "//runtime/src/iree/base/internal:arena",
+        "//runtime/src/iree/testing:gtest",
+        "//runtime/src/iree/testing:gtest_main",
     ],
 )
